Task 8 - Resume Screening Using NLP/requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
numpy
spacy
PyPDF2
python-docx
torch